pyproject.toml
[tool.poetry]
name = "drf-hw"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "drf_hw"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
psycopg2-binary = "^2.9.8"
pillow = "^10.0.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
ipython = "^8.16.1"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
